[fix](inverted index) Multiple code quality improvements
1. Add safety comments for CandidateResult pointer lifetime (ID4)
2. Replace CLuceneError throw with Status return for consistency (ID5)
3. Document shared_lock reentrant behavior in IndexPolicyMgr (ID9)
4. Add thread safety comments for _reader_entries/_key_to_entries (ID10)
5. Use std::get_if instead of std::get to avoid exceptions (ID11)
6. Fix integer type mismatch in loop variable (ID12)

+++ be/src/vec/exprs/vmatch_predicate.cpp
@@ -180,12 +180,12 @@ Status VMatchPredicate::execute_column(VExprContext* context, const Block* block
     })
     ColumnNumbers arguments(_children.size());
     Block temp_block;
-    for (int i = 0; i < _children.size(); ++i) {
+    for (size_t i = 0; i < _children.size(); ++i) {
         ColumnPtr arg_column;
         RETURN_IF_ERROR(_children[i]->execute_column(context, block, count, arg_column));
         auto arg_type = _children[i]->execute_type(block);
         temp_block.insert({arg_column, arg_type, _children[i]->expr_name()});
-        arguments[i] = i;
+        arguments[i] = static_cast<uint32_t>(i);
     }
     uint32_t num_columns_without_result = temp_block.columns();
     // prepare a column to save result
